src/main/java/com/greenfoxacademy/aureuscctribesbackend/security/JWTAuthenticationFilter.java
package com.greenfoxacademy.aureuscctribesbackend.security;

import com.greenfoxacademy.aureuscctribesbackend.services.LoggingServiceImpl;
import io.jsonwebtoken.JwtException;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.authentication.CredentialsExpiredException;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
import org.springframework.web.filter.OncePerRequestFilter;
import javax.servlet.FilterChain;
import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

public class JWTAuthenticationFilter extends OncePerRequestFilter {

  @Autowired
  private JwtService jwtService;
  @Autowired
  private CustomUserDetailsService customUserDetailsService;

  @Autowired
  private LoggingServiceImpl loggingService;

  public JWTAuthenticationFilter() {

  }

  @Override
  protected void doFilterInternal(
      HttpServletRequest request,
      HttpServletResponse response,
      FilterChain filterChain) throws ServletException, IOException {

    String authHeader = request.getHeader("Authorization");

    if (authHeader == null || !authHeader.startsWith("Bearer ")) {
      filterChain.doFilter(request, response);
      return;
    }

    try {
      String jwt = authHeader.substring(7);
      String username = jwtService.extractUsername(jwt);
      UserDetails userDetails;

      if (username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
        userDetails = customUserDetailsService.loadUserByUsername(username);
        if (jwtService.isTokenValid(jwt, userDetails)) {
          UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
              userDetails,
              null,
              userDetails.getAuthorities());
          authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
          SecurityContextHolder.getContext().setAuthentication(authToken);
          filterChain.doFilter(request, response);
        }
      }
    } catch (JwtException e) {
      throw new CredentialsExpiredException(e.getMessage());
    }
  }
}